Update MacOS implementation and refactor POSIX USB CDC Transport
